+++ src/pages/dashboardRequest/components/DashboardRequestActions.tsx
@@ -4,19 +4,25 @@ import { useTranslation, Trans } from 'react-i18next';
 import { Dispatch, SetStateAction, useState } from 'react';
 import { SessionModal, useErrorDispatcher, useLoading } from '@pagopa/selfcare-common-frontend/lib';
 import FileDownloadIcon from '@mui/icons-material/FileDownload';
-import { storageTokenOps } from '@pagopa/selfcare-common-frontend/lib/utils/storage';
+import { AppError } from '@pagopa/selfcare-common-frontend/lib/model/AppError';
 import { LOADING_RETRIEVE_ONBOARDING_REQUEST } from '../../../utils/constants';
 import {
   approveOnboardingPspRequest,
   rejectOnboardingRequest,
 } from '../../../services/onboardingRequestService';
-import { ENV } from '../../../utils/env';
 
 type Props = {
   setShowRejectPage: Dispatch<SetStateAction<boolean | undefined>>;
   setShowConfirmPage: Dispatch<SetStateAction<boolean | undefined>>;
   isToBeValidatedRequest: boolean;
   isGPU: boolean;
+  downloadAttachment?: (
+    setLoadingRetrieveOnboardingRequest: (loading: boolean) => void,
+    addError: (error: AppError) => void,
+    reason?: string,
+    retrieveTokenIdFromUrl?: string,
+    attatchmentName?: string
+  ) => void;
   retrieveTokenIdFromUrl?: string;
   partyName?: string;
   productTitle?: string;
@@ -28,6 +34,7 @@ export default function DashboardRequestActions({
   setShowConfirmPage,
   isToBeValidatedRequest,
   isGPU,
+  downloadAttachment,
   retrieveTokenIdFromUrl,
   partyName,
   productTitle,
@@ -58,30 +65,6 @@ export default function DashboardRequestActions({
     }
   };
 
-  const fileFromReader = async (
-    reader: ReadableStreamDefaultReader<Uint8Array> | undefined
-  ): Promise<string> => {
-    const stream = new ReadableStream({
-      start(controller) {
-        return pump();
-        function pump(): Promise<any> | undefined {
-          return reader?.read().then(({ done, value }) => {
-            if (done) {
-              controller.close();
-              return;
-            }
-            controller.enqueue(value);
-            return pump();
-          });
-        }
-      },
-    });
-    const response = new Response(stream);
-
-    const blob = await response.blob();
-    return URL.createObjectURL(blob);
-  };
-
   const rejectOnboarding = () => {
     setLoadingRetrieveOnboardingRequest(true);
     if (retrieveTokenIdFromUrl) {
@@ -93,55 +76,6 @@ export default function DashboardRequestActions({
         .catch((reason: any) => {
           addError({
             id: `Onboarding request with tokenId: ${retrieveTokenIdFromUrl} not rejected`,
-            blocking: false,
-            techDescription: reason,
-            toNotify: false,
-            error: new Error('INVALID_TOKEN_ID'),
-          });
-        })
-        .finally(() => setLoadingRetrieveOnboardingRequest(false));
-    }
-  };
-
-  const downloadAttachment = () => {
-    setLoadingRetrieveOnboardingRequest(true);
-    const sessionToken = storageTokenOps.read();
-    if (retrieveTokenIdFromUrl) {
-      void fetch(
-        ENV.URL_API.API_ONBOARDING_V2 +
-          `/v2/tokens/${retrieveTokenIdFromUrl}/attachment?=name${attatchmentName}`,
-        {
-          headers: {
-            accept: '*/*',
-            'accept-language': 'it-IT,it;q=0.9,en-US;q=0.8,en;q=0.7',
-            authorization: `Bearer ${sessionToken}`,
-            'content-type': 'application/octet-stream',
-          },
-          method: 'GET',
-        }
-      )
-        .then((response) => {
-          const contentDisposition = response.headers.get('content-disposition');
-          const matchedIndex = contentDisposition?.indexOf('=') as number;
-          const fileName =
-            contentDisposition?.substring(matchedIndex + 1) ?? 'checklist_adesione_gpu.pdf';
-          return response.blob().then((blob) => {
-            const reader = blob.stream().getReader();
-            void fileFromReader(reader).then((url) => {
-              const link = document.createElement('a');
-              // eslint-disable-next-line functional/immutable-data
-              link.href = url;
-              // eslint-disable-next-line functional/immutable-data
-              link.download = fileName;
-              document.body.appendChild(link);
-              link.click();
-            });
-          });
-        })
-        // eslint-disable-next-line sonarjs/no-identical-functions
-        .catch(() => {
-          addError({
-            id: `Onboarding request with tokenId: ${retrieveTokenIdFromUrl} not approved`,
             blocking: false,
             techDescription: reason,
             toNotify: false,
@@ -173,7 +107,16 @@ export default function DashboardRequestActions({
                 variant="naked"
                 sx={{ marginLeft: 3 }}
                 startIcon={<FileDownloadIcon />}
-                onClick={downloadAttachment}
+                onClick={() =>
+                  downloadAttachment &&
+                  downloadAttachment(
+                    setLoadingRetrieveOnboardingRequest,
+                    addError,
+                    reason,
+                    retrieveTokenIdFromUrl,
+                    attatchmentName
+                  )
+                }
               >
                 {t('onboardingRequestPage.actions.downloadButton')}
               </Button>
